Significan't rework of mapping to support the fact the dirs can't be removed.

+++ calibre-web-automated/ha-config/functions.sh
@@ -44,6 +44,14 @@ get_config_value() {
     printf '%s' "${config_value}"
 }
 
+# Function to set an environment variable for later s6 scripts
+set_env_var() {
+    local env_name="${1}"
+    local env_value="${2}"
+
+    printf '%s' "${env_value}" > "/var/run/s6/container_environment/${env_name}"
+}
+
 # Function to set environment variables based on add-on configuration
 set_env_var_from_config() {
     local config_key="${1}"
@@ -55,20 +63,32 @@ set_env_var_from_config() {
         bashio::log.debug "Skipping setting environment variable for empty ${config_key}"
         return
     fi
-    printf '%s' "${config_value}" > "/var/run/s6/container_environment/${env_name}"
+    set_env_var "${env_name}" "${config_value}"
 }
 
 # Function to resolve a path from the add-on configuration and ensure it is a valid absolute path
 resolve_path() {
     local config_key="$1"
+    local default_path="$2"
     local config_value
     local resolved_path
 
     config_value=$(get_config_value "${config_key}")
     if [ -z "${config_value}" ]; then
-        bashio::log.warning "No path configured for ${config_key}, skipping resolution"
-        return
+        if [ -n "${default_path}" ]; then
+            bashio::log.info "No path configured for ${config_key}, using default: ${default_path}"
+            config_value="${default_path}"
+        else
+            bashio::log.warning "No path configured for ${config_key}"
+            return
+        fi
     fi
+    # Ensure the host-mapped directory exists (e.g., /share/my_folder)
+    if ! bashio::fs.directory_exists "${config_value}"; then
+        bashio::log.info "Creating directory ${config_value}"
+        mkdir -p "${config_value}"
+    fi
+    # Resolve the absolute path inside the container
     resolved_path=$(realpath "/${config_value}") || {
         bashio::log.fatal "Failed to resolve ${config_key}: ${config_value}"
         bashio::exit.nok
@@ -81,23 +101,24 @@ resolve_path() {
 map_path() {
     local mapped_path="${1}"
     local app_path="${2}"
+    local file_name="${3}"
 
-    # Ensure the host-mapped directory exists (e.g., /share/my_folder)
-    if ! bashio::fs.directory_exists "${mapped_path}"; then
-        bashio::log.info "Creating directory ${mapped_path}"
-        mkdir -p "${mapped_path}"
-    fi
     # If the mapped path is the same as the app path, no need to create a symlink
     if [ -z "${mapped_path}" ] || [ "${mapped_path}" == "${app_path}" ]; then
         bashio::log.debug "Mapped path ${mapped_path} is the same as app path ${app_path}, skipping symlink creation"
         return
     fi
-    # If the app's default directory exists and isn't a link, move it or remove it
-    if [ -d "${app_path}" ] && [ ! -L "${app_path}" ]; then
+    # If we don't have a filename and the app's default directory exists and isn't a link, move it or remove it
+    if [ -z "${file_name}" ] && [ -d "${app_path}" ] && [ ! -L "${app_path}" ]; then
         bashio::log.warning "Removing existing internal directory to make room for symlink"
         rm -rf "${app_path}"
     fi
     # Create the link
-    ln -s "${mapped_path}" "${app_path}"
-    bashio::log.success "Successfully mapped ${app_path} to ${mapped_path}"
+    if [ -n "${file_name}" ]; then
+        ln -s "${mapped_path}/${file_name}" "${app_path}/${file_name}"
+        bashio::log.success "Successfully mapped ${app_path}/${file_name} to ${mapped_path}/${file_name}"
+    else
+        ln -s "${mapped_path}" "${app_path}"
+        bashio::log.success "Successfully mapped ${app_path} to ${mapped_path}"
+    fi
 }
